001: Faster C, C++, Go, Rust solutions using u32 integer

+++ problems/001/001.go
@@ -6,10 +6,10 @@ import "os"
 import "strconv"
 import "time"
 
-func solution() int {
-    sum := 0
+func solution() uint32 {
+    sum := uint32(0)
 
-    for i := 0; i < 1000; i++ {
+    for i := uint32(0); i < 1000; i++ {
         if i % 3 == 0 || i % 5 == 0 {
             sum += i
         }
